Major change: Most of the visual styling can no longer be done through code, all image, hovering, active, focus, etc must now be configured through the DefaultForms.ini file.
Active, Focus, Hover, Global, Default, and Hidden styles can now be configured and inherited through the .ini file.

Example:
; Class chain: FComponent->FLabel->FButton->FTabButton
;	begin object name=oPlayButton class=FTabButton
;		...
;		StyleNames.Add(NavButton) // Add NavButton as class. You may add more classes to distinguish styles per button, see below QuitButton.
;	end object
[(Standard)Button-TabButton-NavButton FStyle]
ImageColor=(R=255,G=255,B=0,A=255)

[(Standard)Button-TabButton-NavButton:hover FStyle]
ImageColor=(R=240,G=240,B=240,A=255)

The above example shows an example of a PlayButton being created with the assigned style called NavButton which inherits the style from TabButton and Button which are also initialized in the .ini file!

Beware though this is still experimental, does have its issues and still a lot more of the styling needs to be moved to the ini.

[Add]: FCheckBoxStyle.
[Add]: FormsExample as a submodule.